+++ devel/xtf/xtf.h
@@ -372,12 +372,13 @@ namespace xtf {
 			return BOM;
 		}
 		//f Extract and return next character in flow.
-		flw::byte__ Get( utf__ &UTF )
+		flw::byte__ Get( utf__ *UTF = NULL )
 		{
 			if ( !_PrefetchUTF() )
 				qRFwk();
 
-			UTF = _UTF;
+			if ( UTF != NULL )
+				*UTF = _UTF;
 
 //			_F().Skip( _UTF.Size );
 
@@ -417,6 +418,10 @@ namespace xtf {
 
 			return C;
 		}
+		flw::byte__ Get( utf__ &UTF )
+		{
+			return Get( &UTF );
+		}
 		//f NOTA : if '.Line' == 0; a '\n' or a '\r' was unget()'.
 		const pos__ &Position( void ) const
 		{
@@ -442,7 +447,7 @@ namespace xtf {
 		}
 		//f Return the next character in the flow, but let it in the flow.
 		flw::byte__ View(
-			utf__ &UTF,
+			utf__ *UTF = NULL,
 			bso::bool__ HandleNL = false )
 		{
 			if ( !_PrefetchUTF() )
@@ -466,12 +471,18 @@ namespace xtf {
 				}
 			}
 
-			UTF = _UTF;
+			if ( UTF != NULL )
+				*UTF = _UTF;
 
 			return C;
 		}
-		//f True if at end of text.
-		bso::bool__ EndOfFlow( error__ &Error )	// Si erreur, 'ErrorMeaning' est initialis, sinon reste vide.
+		flw::byte__ View(
+			utf__ &UTF,
+			bso::bool__ HandleNL = false )
+		{
+			return View( &UTF, HandleNL );
+		}
+		bso::bool__ EndOfFlow( error__ *Error = NULL )	// Si erreur, 'ErrorMeaning' est initialis, sinon reste vide.
 		{ 
 			if ( _Error == e_NoError ) {
 				if ( _UTF.Size != 0 )
@@ -481,15 +492,25 @@ namespace xtf {
 					return true;
 
 				if ( !_PrefetchUTF() ) {
-					Error = _Error = eEncodingDiscrepancy; 
+					_Error = eEncodingDiscrepancy; 
+
+					if ( Error != NULL )
+						*Error = _Error;
+
 					return true;
 				}
 
 				return false;
 			} else {
-				Error = _Error;
+				if ( Error != NULL )
+					*Error = _Error;
+
 				return true;
 			}
+		}
+		bso::bool__ EndOfFlow( error__ &Error )
+		{
+			return EndOfFlow( &Error );
 		}
 		void Dismiss( void )
 		{