+++ stable/sclrgstry.h
@@ -41,6 +41,7 @@ namespace sclrgstry {
 
 	using rgstry::tags_;
 	using rgstry::tags;
+	using rgstry::rEntry;
 
 	typedef rgstry::multi_level_registry_ registry_;