+++ tools/rncq/CLI/rncq.h
@@ -1,0 +1,40 @@
+/*
+	Copyright (C) 2016 by Claude SIMON (http://zeusw.org/epeios/contact.html).
+
+	This file is part of 'RNCq'.
+
+    'RNCq' is free software: you can redistribute it and/or modify it
+    under the terms of the GNU Affero General Public License as published
+    by the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    'RNCq' is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with 'RNCq'.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef RNCQ_INC_
+# define RNCQ_INC_
+
+# define BEGIN
+// Don't remove what between above 'BEGIN' and below 'END'.
+/* PLACEHOLDER */
+
+# define NAME_MC			"RNCq"
+# define NAME_UC			"RNCQ"
+# define NAME_LC			"rncq"
+# define WEBSITE_URL		"http://q37.info/"
+# define AUTHOR_NAME		"Claude SIMON"
+# define AUTHOR_CONTACT		"http://q37.info/contact/"
+# define OWNER_NAME			"Claude SIMON"
+# define OWNER_CONTACT		"http://q37.info/contact/"
+# define COPYRIGHT			COPYRIGHT_YEARS " " OWNER_NAME " (" OWNER_CONTACT ")"	
+
+# define END
+
+#endif
+